Added docs output to stencil config
This autogenerates documentation upon build that is useful to figure out dependencies and onboarding people on stencil projects. In this case it is to help me determine quicker which props I can remove safely as props should not be changed from within the same component during rendering and this is happening a lot right now in the browser console warnings. It is a performance risk and could lead to infinite loops in some scenarios.