Stored menu items in session storage.
Because menu items don't change that often, storing them for the session makes sense to save on web and server resources.
This only persists for a client session, which means if the user closes the browser and repopens it, the user will get wathever new/updated menu items available on first visit.